+++ Assets/RPG System/Scripts/Combat/CombatEffects/MagicDamage.cs
@@ -9,10 +9,10 @@ public class MagicDamage : CombatEffectBase {
     {
     }
 
-    public override void ApplyEffect(float damage)
+    public override void ApplyEffect(int damage)
     {
         // TODO: Calculate damage using effectExpression
-        target.entity.hitPoints -= (int)damage;
+        target.entity.hitPoints -= damage;
     }
 
     public override bool IsValid
