+++ Assets/RPG System/Scripts/Combat/CombatEffects/ManaDrain.cs
@@ -12,9 +12,9 @@ public class ManaDrain : CombatEffectBase {
         this.drain = drain;
     }
 
-    public override void ApplyEffect(float drain)
+    public override void ApplyEffect(int drain)
     {
-        target.entity.mana -= (int)drain;
+        target.entity.mana -= drain;
     }
 
     public override bool IsValid
